requirements.txt
requests>=2.31.0
anthropic>=0.25.0
pyyaml>=6.0
